pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xpra-ai-control"
version = "0.1.0"
description = "AI Control Layer for X11 via Xpra"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyzmq>=25.0",
    "msgpack>=1.0",
    "Pillow>=10.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
llm = [
    "httpx>=0.25",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
xpra-ai-agent = "xpra_ai_control.agent:main"

[tool.setuptools.packages.find]
include = ["xpra_ai_control*"]

[tool.ruff]
line-length = 100
target-version = "py310"
